server/requirements.txt
aioconsole==0.7.0
annotated-types==0.6.0
asttokens==2.4.1
colorama==0.4.6
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
executing==2.0.1
fonttools==4.51.0
icecream==2.1.3
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pydantic_core==2.18.2
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
typing_extensions==4.11.0
websockets==12.0
wordcloud==1.9.3
